% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guppies}
\alias{guppies}
\title{# Spot brightness of male guppies}
\format{A data frame with 25036 objects of 5 variables.
\describe{
  \item{Spot.brightness}{Number: scale from 1-20 (dull to bright)}
  \item{Day}{Number: Day of experiment}
  \item{Substrate}{Factor: substrate in the tank}
  \item{Predator}{Factor: predator treatment for tank}
  \item{Predator.number}{Number: predators added to tank}
}}
\usage{
guppies

guppies
}
\description{
## guppies

Guppies data converted into frequency counts, grouped by other variables. Each row represents a number of male guppies with the same spot brightness from a populations with the same conditions, on the same day of the experiment.
}
\details{
A dataset containing the spot brightness of male guppies along with other attributes related to the treatment they were exposed to during virtual SimBio experiments. Each row represents one male guppy.
}
\keyword{datasets}
