% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guppies_extra}
\alias{guppies_extra}
\title{## guppies_extra}
\format{A data frame with 2849 rows and 7 variables:
\describe{
  \item{Spot.brightness}{Number: scale from 1-20 (dull to bright)}
  \item{Day}{Number: Day of experiment}
  \item{Substrate}{Factor: substrate in the tank}
  \item{Predator}{Factor: predator species added to tank}
  \item{Predator.number}{Number: predators added to tank}
  \item{Additional.R.hartii}{Number: *R. hartii* added to tank alongside any other predators}
  \item{Starting.food}{Number: food items in tank at start of experiment}
}}
\usage{
guppies_extra
}
\description{
Additional conditions beyond the main experiment. Each row represents one male guppy.
}
\keyword{datasets}
