% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_density}
\alias{predict_density}
\title{Get predictions from predator density model}
\usage{
predict_density(x = density, predator, mod.object = fit_density)
}
\arguments{
\item{x}{Vector of new x values to make predictions for}

\item{predator}{Character string definig predator factor to make predictions for.}

\item{mod.object}{the model object to use for these predictions.}
}
\value{
A data frame of predcitions with confidence intervals.
}
\description{
Takes a model object and gathers predictions for a specific predator factor.
}
