% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_substrate}
\alias{predict_substrate}
\title{Get predictions from substrate interaction model}
\usage{
predict_substrate(x = density, predator, substrate,
  mod.object = fit_substrate_density)
}
\arguments{
\item{x}{Vector of new x values to make predictions for}

\item{predator}{Character string definig predator factor to make predictions for.}

\item{substrate}{Character string defining substrate to make predictions for.}

\item{mod.object}{the model object to use for these predictions.}
}
\value{
A data frame of predcitions with confidence intervals.
}
\description{
Takes a model object and gathers predictions for specific predator and substrate factors.
}
