% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rgba}
\alias{rgba}
\title{Change the alpha value of colours}
\source{
https://magesblog.com/post/2013-04-30-how-to-change-alpha-value-of-colours-in/
}
\usage{
myColours = c(1, "steelblue", "#FFBB00", rgb(0.4, 0.2, 0.3))
myColoursAlpha <- rgba(myColours, alpha=0.4)
}
\arguments{
\item{col}{a vector of colours}

\item{alpha}{alpha valuses range from 0-1, 0 being fully transparent and 1 being fully opaque.}
}
\value{
a vector of alpha colours (notice longer codes)
}
\description{
Reduce the alpha value (level of transparency) of colours to identify patterns of over-plotting when displaying lots of data points.
}
