% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{roses_set_ggtheme}
\alias{roses_set_ggtheme}
\title{Rose's favourite ggplot theme settings}
\usage{
plot_object + roses_ggtheme()
plot_object + roses_ggtheme(legend.position = "none")
plot_object + roses_ggtheme(legend.title = element_blank())
}
\arguments{
\item{...}{Accepts all the usual theme() params except those already set bby default.}
}
\description{
Rose's favourite ggplot theme settings
}
