% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{roses_unicode}
\alias{roses_unicode}
\title{Rose's favourite unicodes}
\usage{
roses_unicode(shorthand)
}
\arguments{
\item{shorthand}{my nickname for the unicaode character}
}
\value{
hexidecimal form of the unicode, defaults to filled dot.
}
\description{
These are all tried and tested as pch vals. Unicode characters work really nicely as pch vals with alpha transparency as they are one solid vector layer - the whole icon will be the same transparency and and will have nice high res smooth edges.
}
